refactor: address maintainability audit findings (#198)
- Split the Reels wizard out of schedule_instagram_posts.py into its
  own planning/instagram/instagram_reels.py, mirroring the LinkedIn
  module split.
- Split NotionNewsletterBuilder's Notion-read, grouping, and
  HTML-render concerns in build_newsletter.py so substack_draft.py can
  reuse grouping without importing the whole class.
- Replace the hardcoded Substack-platform string check in
  data_processor.py with a mapping.json-driven
  optional_if_first_record flag.
